+++ src/main/kotlin/dev/mizarc/waystonewarps/interaction/menus/management/WarpManagementMenu.kt
@@ -57,7 +57,7 @@ class WarpManagementMenu(private val player: Player, private val menuNavigator: 
                 toggleLock.execute(
                     playerId = player.uniqueId,
                     warpId = warp.id,
-                    bypassOwnership = player.hasPermission("waystonewarps.bypass.accesscontrol"),
+                    bypassOwnership = player.hasPermission("waystonewarps.bypass.access_control"),
                 )
                 open()
             }